% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_items.R
\docType{class}
\name{all_items-class}
\alias{all_items-class}
\title{All Items in an order}
\description{
\emph{all_items} is an S4 class to represent all of the items in an order from a business owner's perspective.
}
\section{Slots}{

\describe{
\item{\code{items}}{A list of \emph{single_item} objects.}
}}

\examples{
\dontrun{shopping_cart <- new('all_items', items = list(items, item2))}

}
