% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\docType{methods}
\name{discount}
\alias{discount}
\alias{discount,single_item-method}
\title{Discount of an item}
\usage{
discount(obj, ...)

\S4method{discount}{single_item}(obj, ...)
}
\arguments{
\item{obj}{An object.}

\item{item}{An item to be added.}
}
\value{
The remaining fraction a value in decimal after the single item is discounted.
}
\description{
\emph{discount} returns the remaining fraction a value after it is discounted.
}
