% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\docType{methods}
\name{display}
\alias{display}
\alias{display,all_items-method}
\title{Display Content}
\usage{
display(obj, ...)

\S4method{display}{all_items}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\value{
A data frame describing the object. # for \emph{all_items} objects
}
\description{
Similar to \emph{print}, but \emph{display} shows the content of an object in a more viewer-friendly format while \emph{print} provides a compact form of the content for bookkeeping.
}
