% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate.R
\docType{methods}
\name{find_duplicate,product_logs,character,character,character-method}
\alias{find_duplicate,product_logs,character,character,character-method}
\title{Check if an order number has a duplicate in the database}
\usage{
\S4method{find_duplicate}{product_logs,character,character,character}(obj, date,
  orn, sup, ...)
}
\arguments{
\item{date}{A character string representing the date of the transaction.}

\item{orn}{A character string representing the order number of the transaction.}

\item{sup}{A character string representing the supplier of the transaction.}
}
\value{
The location of the duplicate if one exists and character(0) otherwise.
}
\description{
\emph{find_duplicate} checks if an order number has a duplicate in the database and returns the location of the duplicate if one exists.
}
\examples{
\dontrun{find_duplicate(a_log, '2016-12-23', 'A123456', 'Amazon')}
}
