% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice.R
\docType{class}
\name{invoice-class}
\alias{invoice-class}
\title{Invoice for an order}
\description{
An \emph{invoice} object handles order number generation and keeps the order numbers in a tracking system.
}
\section{Slots}{

\describe{
\item{\code{fname}}{A character string indicating the name of the file in which thetracking system is stored.}

\item{\code{part1}}{A character string describing the first part of the invoice number pattern.}

\item{\code{part2}}{A character string describing the second part of the invoice number pattern.}

\item{\code{count_begins_at}}{A character string indicating the number at which the count begins and the number of placeholders for the count.}
}}

\examples{
\dontrun{
vt <- new('invoice', fname = 'invoice_test', part1 = 'abc', part2 = 'xyz', count_begins_at = '3146')
}

}
