% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice.R
\docType{methods}
\name{invoice.demo,invoice-method}
\alias{invoice.demo,invoice-method}
\title{Demo for invoice generator for orders}
\usage{
\S4method{invoice.demo}{invoice}(obj, is_sampling, inv)
}
\arguments{
\item{obj}{The invoice object.}

\item{is_sampling}{A boolean value indicating whether the transaction is in fact a sampling for the production or not.}

\item{inv_df}{A data frame containing a log of the invoice generated.}
}
\value{
A list containing a character string representation of the order number if it is indeed a sale and 'N/A' otherwise and an undated log of generated invoices.
}
\description{
The \emph{invoice.demo} method generates an order number for the transaction when a sale is made. An order number generated has a pattern consists of four parts: A customized first part, followed by the 2-digit represention of the year when the order number is generated, follow by a customized second part, and completed by the (possibly encryted) count of order number issued. *At the beginnig of the following years, the count will be reset to the \emph{count_begins_at} assigned at the construction of the object, with the 2-digit year modified.
}
\examples{
\dontrun{invoice.demo(vt, FALSE)}

}
