% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice.R
\docType{methods}
\name{invoice,invoice-method}
\alias{invoice,invoice-method}
\title{Invoice generator for orders}
\usage{
\S4method{invoice}{invoice}(obj, is_sampling)
}
\arguments{
\item{obj}{The invoice object.}

\item{is_sampling}{A boolean value indicating whether the transaction is in fact a sampling for the production or not.}
}
\value{
A character string representation of the order number if it is indeed a sale and 'N/A' otherwise.
}
\description{
The \emph{invoice} method generates an order number for the transaction when a sale is made. An order number generated has a pattern consists of four parts: A customized first part, followed by the 2-digit represention of the year when the order number is generated, follow by a customized second part, and completed by the (possibly encryted) count of order number issued. *At the beginnig of the following years, the count will be reset to the \emph{count_begins_at} assigned at the construction of the object, with the 2-digit year modified.
}
\examples{
\dontrun{invoice(vt, FALSE)}

}
