% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\docType{methods}
\name{is_empty}
\alias{is_empty}
\alias{is_empty,single_item-method}
\alias{is_empty,all_items-method}
\alias{is_empty,product_transac-method}
\alias{is_empty,product_logs-method}
\title{Check if the Object Is Empty}
\usage{
is_empty(obj, ...)

\S4method{is_empty}{single_item}(obj)

\S4method{is_empty}{all_items}(obj)

\S4method{is_empty}{product_transac}(obj)

\S4method{is_empty}{product_logs}(obj)
}
\arguments{
\item{obj}{An object}
}
\value{
A boolean value describing whether the object has any nontrivial slots.

A boolean value describing whether the object has any nontrivial slots.

A boolean value describing whether the object has any nontrivial slots.

A boolean value describing whether the object has any nontrivial slots.
}
\description{
\emph{is_empty} checks if the object has only trivial slots and returns a boolean value for the result.
}
