% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\docType{methods}
\name{join,product_logs,product_logs-method}
\alias{join,product_logs,product_logs-method}
\title{Append another log to a log}
\usage{
\S4method{join}{product_logs,product_logs}(obj, another, ...)
}
\arguments{
\item{another}{A product_logs object whose entries to be added to the main argument.}
}
\value{
A new \emph{product_logs} object with the combined logs.
}
\description{
\emph{join} appends another log to the calling object and returns a new \emph{product_logs} object with the combined logs.
}
\examples{
\dontrun{join(log1, log2)}
}
