% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_track.R
\docType{methods}
\name{order_track,product_logs-method}
\alias{order_track,product_logs-method}
\title{Create codes for the order numbers}
\usage{
\S4method{order_track}{product_logs}(obj, ...)
}
\value{
A vector of character strings, the new codes for the order number for each transaction in the log.
}
\description{
\emph{order_track} creates a new code for the order number for each transaction in the log.
}
\examples{
\dontrun{order_track(a_log)}
}
