% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,single_item-method}
\alias{print,all_items-method}
\alias{print,product_transac-method}
\alias{print,product_logs-method}
\title{Print Objects}
\usage{
print(obj, ...)

\S4method{print}{single_item}(obj, ...)

\S4method{print}{all_items}(obj, ...)

\S4method{print}{product_transac}(obj, ...)

\S4method{print}{product_logs}(obj, ...)
}
\arguments{
\item{obj}{An object}
}
\value{
A character string describing the object. # for \emph{single_item} objects

A character string describing the object. # for \emph{all_items} objects

A data frame describing the object. # for \emph{product_logs} objects

A data frame describing the object. # for \emph{product_logs} objects
}
\description{
Rather than printing out its argument, \emph{print} siliently returns a representation of its argument in various forms such as character string and data frame.
}
