% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_logs.R
\docType{class}
\name{product_logs-class}
\alias{product_logs-class}
\title{PRODUCT LOGS}
\description{
\emph{product_logs} is an S4 class to represent a business transaction log.
}
\section{Slots}{

\describe{
\item{\code{logs}}{A list of \emph{product_transac} objects}
}}

\examples{
\dontrun{a_log <- new('product_logs', logs = list(product_transaction1, product_transaction2))}

}
