% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_last.R
\docType{methods}
\name{remove_last}
\alias{remove_last}
\alias{remove_last,all_items-method}
\title{Remove the most last element}
\usage{
remove_last(obj, ...)

\S4method{remove_last}{all_items}(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\description{
\emph{remove_last} removes the most recently added item from a structure within the object.
}
