% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_items_info.R
\docType{methods}
\name{show_all_items_info,product_logs-method}
\alias{show_all_items_info,product_logs-method}
\title{Show information of all items in a log for summary}
\usage{
\S4method{show_all_items_info}{product_logs}(obj, ...)
}
\value{
A data frame containing informations about each item inside the log.
}
\description{
\emph{show_all_items_info} extracts informations about each item including the date of transaction, type of transaction, category, and model inside the log for later summary.
}
\examples{
\dontrun{show_all_items_info(a_log)}
}
