% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_duplicate.R
\docType{methods}
\name{show_duplicate,product_logs,character,character,character-method}
\alias{show_duplicate,product_logs,character,character,character-method}
\title{Display the record with the same order number}
\usage{
\S4method{show_duplicate}{product_logs,character,character,character}(obj, date,
  orn, sup, ...)
}
\arguments{
\item{date}{A character string representing the date of the transaction.}

\item{orn}{A character string representing the order number of the transaction.}

\item{sup}{A character string representing the supplier of the transaction.}
}
\value{
The existing record of transaction in a data frame if there is a duplicate and an empty data frame otherwise.
}
\description{
\emph{"show_duplicate"} displays the existing record of transaction in the log who has the same order number as the one entered by the user.
}
\examples{
\dontrun{show_duplicate(a_log, '2016-12-23', 'A123456', 'Amazon')}
}
