% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_stock.R
\docType{methods}
\name{show_stock}
\alias{show_stock}
\alias{show_stock,product_logs-method}
\title{Show the stock of products}
\usage{
show_stock(obj, ...)

\S4method{show_stock}{product_logs}(obj, ...)
}
\arguments{
\item{obj}{An object.}
}
\value{
The result of the extraction.
}
\description{
\emph{show_stock} extracts the pertinent items' information for each item from each transaction record of a business log and returns the result for further calculation of the stock status of product.
}
\examples{
\dontrun{show_stock(business_logs)}
}
