% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_item.R
\docType{class}
\name{single_item-class}
\alias{single_item-class}
\title{Single Item in an order}
\description{
\emph{single_item} is an S4 class to represent an item in an order from a business owner's perspective.
}
\section{Slots}{

\describe{
\item{\code{category}}{A character string describing the category of the item.}

\item{\code{descrp}}{A character string describing the item.}

\item{\code{model}}{A character string representing the model of the item.}

\item{\code{quant}}{A numeric value representing the quantity of the item in the order.}

\item{\code{price}}{A numeric value representing the unit price of the item.}

\item{\code{discount}}{A character string representing the discount in percentage with the '\%' symbol on the item.}

\item{\code{amount}}{A numeric value representing the amount due for the item at the given unit price, quantity, and discount.}

\item{\code{sample}}{A character string describing whether the item is a sample for a customer or not with values 'Yes' and 'No'.}

\item{\code{category}}{A character string describing whether the item is a sampling item for the production or not with values 'Yes' and 'No'.}
}}

\examples{
\dontrun{

an_item <- new('single_item', category = 'some category', descrp = 'some description', 

model = 'some model', quant = 1, price = 10, discount = '10\%', amount = 9, sample = 'No', 

sampling = 'No')

}


}
